Add files via upload
This repository contains code and analysis tools developed as part of the CERN Summer Student Program 2024. The project focuses on the analysis of particle collision data, specifically targeting the reconstruction and efficiency evaluation of electrons and muons from tau decay events using the ROOT framework. The tools provided facilitate the visualization and interpretation of particle physics data, contributing to ongoing research in high-energy physics, in the charged lepton flavor number violation (CLFV) in the tauon decay into (muon + 2 electrons), (2 muons+ electron) and (3 electrons)

    The code is structured to analyze particle collision data, focusing on the reconstruction of electrons and muons from generated particles.
    It effectively uses ROOT's capabilities to create histograms and calculate efficiencies, which are essential for particle physics research.
    The modular design with functions allows for easy maintenance and potential expansion of the analysis.